ioctl request type
fix keyp on 64-bit architectures

Bug-Debian: https://bugs.debian.org/744802

Gbp-Pq: Name fix_keyp_64bit.patch

+++ files.c
@@ -701,7 +701,7 @@ NODE *leofp(NODE *args) {
 
 NODE *lkeyp(NODE *args) {
 #if defined(unix) | defined(__WXMSW__)
-    long nc;
+    int nc;
 #endif
     int c;
 #ifdef WIN32
